添加回复功能的异常处理

+++ src/main/java/com/pigman/community/controller/PublishController.java
@@ -35,7 +35,7 @@ public class PublishController {
     public String doPublish(@RequestParam( value="title",required=false) String title,
                             @RequestParam(value="description",required=false) String description,
                             @RequestParam(value="tag",required=false)String tag,
-                            @RequestParam(value="id",required=false) Integer id,
+                            @RequestParam(value="id",required=false) Long id,
                             HttpServletRequest request,
                             Model model){
 
@@ -91,7 +91,7 @@ public class PublishController {
      * @return
      */
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name="id")Integer id,
+    public String edit(@PathVariable(name="id")Long id,
                        Model model){
         QuestionDTO question = questionService.findById(id);
         model.addAttribute("title",question.getTitle());
