+++ src/main/java/com/pigman/community/controller/QuestionController.java
@@ -16,7 +16,7 @@ public class QuestionController {
 
     @GetMapping("/question/{id}")
     public String question(
-            @PathVariable(name="id")Integer id,
+            @PathVariable(name="id")Long id,
             Model model){
         QuestionDTO questionDTO = questionService.findById(id);
         //累加评论
